pyproject.toml
[project]
name = "Compiladores"
version = "0.1.0"
description = ""
authors = [
    {name = "Luis Enilton", email = "luiseniltonxy@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "antlr4-python3-runtime>=4.13.2,<5.0.0"
]

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
antlr4-python3-runtime = "^4.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
